move fake env vars to test file
